vta_separator_v1.1.py
This will probably be my final change to the VTA separator. Compared to the last version, version 1.1 evenly splits the distance that vertices dead in the middle have to travel when applying two shape keys, for example smile_L and smile_R. Before, vertices with an X location of 0.0 would always be moved with the right side, which causes a sharp deformation in the face. This is no longer the case, and the bridge between separate flexes will now be as smooth as possible.