Update and rename vta_separator_v1.1.2.py to vta_separator_v1.1.3.py
Ok! Pretty sure this is going to be my final update, but no promises. This change makes it so that vertices that were defined in the basis shape key as having an X location of 0, stay in the middle when activating a shape key. In heavy's case, some vertices were moved out of the middle, so they were separated into either side instead of being evenly split between the two sides. Now, they will permanently stay in the middle.